[MIPS] Fix ll/sc after 7dd547e5ab6b31e7a0cfc182d3ad131dd55a948f
After 7dd547e5ab6b31e7a0cfc182d3ad131dd55a948f the env->llval value is
loaded as an unsigned value (instead of sign-extended as before).
Therefore, the CMPXCHG in gen_st_cond() in translate.c fails if the sign
bit is set in the loaded value.
Fix this by sign-extending the llval value for the 32-bit case.

I discovered this issue because FreeBSD MIPS64 was looping forever in an
atomic helper function when trying to start /sbin/init.

Feel free to apply this patch upstream, as I cannot submit the patch
myself due to the complicated contribution process.

Signed-off-by: Alex Richardson <Alexander.Richardson@cl.cam.ac.uk>