Refactor instruction logging pause/resume.
Move to using per-CPU log levels. Each CPU has associated a separate log
strategy (currently ALL, NONE, USER-ONLY), and a pause/resume flag.
The CPU_LOG_INSTR level is the global enable for instruction logging,
this is set at startup and never (normally) unset, we still need the flag to
keep the trace file open.

The pause/resume switching logic is moved to a callback that is run
within the CPU exclusive context at the same time of flushing the TCG
buffer. This makes the switching logic cleaner and the synchronization
is easier to handle. We also are guaranteed to be called after the side-effects
for the current instructions are completely handled by the target.